Use Trachemys genus instead of Trachemys scripta
Only 1 of the 3 Trachemys scripta subspecies are linked to Trachemys
scripta in the GBIF backbone
(https://github.com/gbif/checklistbank/issues/24). That means we miss 2
Trachemys scripta subspecies if we query for Trachemys scripta. They
are all linked to the genus however, so well query on that instead.
There are no native Trachemys in Belgium, so we don’t query to much
either that way.

Since this is supposed to be a temporary patch, I have only updated the
name from species to genus in the verified-checklist (with Trachemys
scripta still in verbatimScientificName) and via the notebook in the
resulting aggregated-checklist. I have not updated the name in the
individual t0 files in raw and interim, or the automatically generated
concatenated-checklist.

The name Trachemys scripta is also listed in Harmonia. I kept it as
such there, which is why in the aggregated checklist there are now 2
records: Trachemys and Trachemys scripta.